add-git-ignore fix-column-type refactor-scheduler-config
