+++ src/main/java/pro/sky/telegrambot/configuration/SchedulerConfig.java
@@ -1,0 +1,39 @@
+package pro.sky.telegrambot.configuration;
+
+import com.pengrad.telegrambot.TelegramBot;
+import com.pengrad.telegrambot.request.SendMessage;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.scheduling.annotation.EnableAsync;
+import org.springframework.scheduling.annotation.EnableScheduling;
+import org.springframework.scheduling.annotation.Scheduled;
+import pro.sky.telegrambot.entity.NotificationTask;
+import pro.sky.telegrambot.repository.NotificationTaskRepository;
+
+import java.time.LocalDateTime;
+import java.time.temporal.ChronoUnit;
+
+
+@Configuration
+@EnableScheduling
+@EnableAsync
+@ConditionalOnProperty(name = "scheduler.enabled", matchIfMissing = true)
+public class SchedulerConfig {
+    private final NotificationTaskRepository repository;
+    private final TelegramBot telegramBot;
+
+
+    public SchedulerConfig(NotificationTaskRepository repository, TelegramBot telegramBot) {
+        this.repository = repository;
+        this.telegramBot = telegramBot;
+    }
+
+    @Scheduled(cron = "0 0/1 * * * *")
+    public void run() {
+        NotificationTask task = repository.findByTimeSend(LocalDateTime.now().truncatedTo(ChronoUnit.MINUTES));
+        if (task != null) {
+            SendMessage reminder = new SendMessage(task.getChatId(), task.getMessage());
+            telegramBot.execute(reminder);
+        }
+    }
+}
