+++ src/main/resources/liquibase/scripts/tasks.sql
@@ -6,4 +6,8 @@ CREATE TABLE notification_task (
     chat_id   BIGINT,
     message   TEXT,
     time_send TIMESTAMP
-)
+);
+
+-- changeset iavdeyev:2
+ALTER TABLE notification_task
+    ALTER COLUMN id TYPE BIGINT USING (id::BIGINT);
